+++ packages/content/src/layout/index.ts
@@ -36,9 +36,9 @@ export { LayoutService, LayoutServiceConfig, GRAPHQL_LAYOUT_QUERY_NAME } from '.
 
 export { getDesignLibraryStylesheetLinks } from './themes';
 
-export { rewriteEdgeHostInResponse, containsDefaultEdgeHost } from './rewrite-edge-host';
-
 export {
-  rewriteContentInLayout,
-  ContentRewriteOptions,
-} from './content-rewrite';
+  rewriteEdgeHostInResponse,
+  containsDefaultEdgeHost,
+  getDefaultMediaUrlTransformer,
+  applyMediaUrlRewrite,
+} from './rewrite-edge-host';